Stop forcing Memory access in createCreep
Previously createCreep always created a memory object for the creep
being created. This meant that even when using RawMemory, the default
JSON parser would be invoked anyways when creating a creep. This is
especially a problem if the raw memory cannot be parsed into a JSON
object, resulting in an exception in createCreep.

With this commit the memory access is only done when the creepMemory
parameter is not undefined.