Move all damage calculations to the end of the tick, and strictly always apply damage before healing.
This makes actual actions correspond much more closely with the "everything is based off of things at the start of the tick" model.

Before this, if a newly-damaged creep worked at full capacity the tick it was damaged was essentially random, depending on whether the attack intention or the work intention 'happened first'.

In addition, relating to healing, previously, it would be essentially random whether or not a heal operation on a fully-healed creep would do anything if the creep was also damaged during the tick. Sometimes, if the attack intention was in the database 'before' the heal intention, the heal would do something, and if the heal was before the attack it would do nothing.

Now, everything will work based on how much health and what bodyparts the creep had at the start of the tick. As well, this treats heal operations as a 'heal for the entire tick', or 'overheal' action - allowing a heal to always heal/counteract some of the damage applied during the same tick if a full-health creep is both attacked and healed.