Expose Game.setIntent
This allows users to use an alternate style of interacting with the game
API, such as a functional approach like:

```javascript
_(Game.creeps).map(creepBrain)
              .reduce(Game.setIntent);
```

Where creepBrain is a function from Creep -> intended action.

Security-wise there shouldn't be a problem with this, since
the latter processing of intents needs to be secure by itself.