feat: Add ability to specify spawn name (#50)
* Add ability to specify spawn name

Right now the only way to specify a spawn name is by using the web
interface, otherwise they are autogenerated. This change adds an option
for specifying the name.

* move `createdSpawnNames` to prevent duplicate custom names

* actually check for reuse of spawn name

* check name against string type, rather than check undefined

* check for construction sites with the provided spawn name

* Add ability to specify spawn name

Right now the only way to specify a spawn name is by using the web
interface, otherwise they are autogenerated. This change adds an option
for specifying the name.

* move `createdSpawnNames` to prevent duplicate custom names

* actually check for reuse of spawn name

* check name against string type, rather than check undefined

* check for construction sites with the provided spawn name