Eliminate many one-element arrays in spatials
In the vast majority of cases, there is only one object of a given type at a given location in a room. Storing these objects directly in the spatial lookup tables rather than creating a new one-element array for every object eliminates most of these arrays while still support multiple objects of a type by detecting that case and accounting for it.

On my local private server running IVM with a small amount of room objects, this change resulted in a ~15% reduction in heap generated by the for loop over roomObjects in makeGameObject.