Add bounds checking for getTerrainAt
Currently, if the user passes an `x,y`-pair into `Game.map.getTerrainAt`,
that are out of bounds, the returned value is confusing:

- If `x` is too large, it will wrap around to next `y` values.
- If `x` is negative, it will wrap around to previous `y` values.
- If `y` is too large or negative, it will return `'plain'`.

Having the return value instead be `undefined` in these cases makes it
clearer to users debugging malfunctioning code what is going on.
Additionally, it returns `false` when compared with `'plain'`, making
searching for free spaces near mining spots a lot easier for new players.