effects gets set to undefined if array length is 0
Right now, there is a discrepency in how the `.effects` property on RoomObjects is handled.

If there are no effects on an object, an object that has never had an effect on it has an undefined `effects` property, whereas objects that have had an effect on them in the past have an `effects` property of `[]`

This change sets `effects` to undefined if there are no effects on the object, making this property consistent for all room objects.