Fix CostMatrix.clone causing unnecessary memory pressure
Since we're about to replace the whole of `_bits` with a new value,
calling the constructor leads to the code creating a full 2500 big
`UInt8Array` just to replace it instantly.

Fix the issue by doing the same trick `CostMatrix.deserialize` does;
use `Object.create` to skip the constructor cost.